+++ udata/core/activity/tasks.py
@@ -1,11 +1,51 @@
 # -*- coding: utf-8 -*-
 from __future__ import unicode_literals
 
-# from celery.utils.log import get_task_logger
-from udata.tasks import celery
+import warnings
+import logging
+
+from udata.models import db, User, Organization
+from udata.tasks import task
+
+from .signals import new_activity
+
+log = logging.getLogger(__name__)
 
 
-@celery.task
+@new_activity.connect
+def delay_activity(cls, related_to, actor, organization=None):
+    emit_activity.delay(
+        cls.__name__,
+        str(actor.id),
+        related_to_cls=related_to.__class__.__name__,
+        related_to_id=str(related_to.id),
+        organization_id=str(organization.id) if organization else None,
+    )
+
+
+@task
+def emit_activity(classname, actor_id, related_to_cls, related_to_id,
+                  organization_id=None):
+    log.debug('Emit new activity: %s %s %s %s %s',
+              classname, actor_id, related_to_cls,
+              related_to_id, organization_id)
+    cls = db.resolve_model(classname)
+    actor = User.objects.get(pk=actor_id)
+    related_to = db.resolve_model(related_to_cls).objects.get(pk=related_to_id)
+    if organization_id:
+        organization = Organization.objects.get(pk=organization_id)
+    else:
+        organization = None
+    cls.objects.create(actor=actor, related_to=related_to,
+                       organization=organization)
+
+
+@task
 def write_activity(cls, actor, related_to, organization=None, **kwargs):
+    warnings.warn(
+        'write_activity task is deprecated and '
+        'will be removed in udata 2.0',
+        DeprecationWarning
+    )
     cls.objects.create(actor=actor, related_to=related_to,
                        organization=organization, kwargs=kwargs)