+++ udata/core/dataset/activities.py
@@ -4,7 +4,6 @@ from __future__ import unicode_literals
 from udata.auth import current_user
 from udata.i18n import lazy_gettext as _
 from udata.models import db, Dataset, Activity
-from udata.core.activity.tasks import write_activity
 
 
 __all__ = (
@@ -42,24 +41,18 @@ class UserDeletedDataset(DatasetRelatedActivity, Activity):
 def on_user_created_dataset(dataset):
     if (not dataset.private and current_user and
             current_user.is_authenticated):
-        user = current_user._get_current_object()
-        organization = dataset.organization
-        write_activity.delay(UserCreatedDataset, user, dataset, organization)
+        UserCreatedDataset.emit(dataset, dataset.organization)
 
 
 @Dataset.on_update.connect
 def on_user_updated_dataset(dataset):
     if (not dataset.private and current_user and
             current_user.is_authenticated):
-        user = current_user._get_current_object()
-        organization = dataset.organization
-        write_activity.delay(UserUpdatedDataset, user, dataset, organization)
+        UserUpdatedDataset.emit(dataset, dataset.organization)
 
 
 @Dataset.on_delete.connect
 def on_user_deleted_dataset(dataset):
     if (not dataset.private and current_user and
             current_user.is_authenticated):
-        user = current_user._get_current_object()
-        organization = dataset.organization
-        write_activity.delay(UserDeletedDataset, user, dataset, organization)
+        UserDeletedDataset.emit(dataset, dataset.organization)