+++ udata/search/__init__.py
@@ -4,6 +4,7 @@ from __future__ import unicode_literals
 import bson
 import datetime
 import logging
+import warnings
 
 from elasticsearch import Elasticsearch
 from elasticsearch.helpers import scan
@@ -15,7 +16,8 @@ from mongoengine.signals import post_save, post_delete
 from speaklater import is_lazy_string
 from werkzeug.local import LocalProxy
 
-from udata.tasks import task
+from udata.models import db
+from udata.tasks import task, as_task_param
 
 
 from . import analysis
@@ -120,12 +122,21 @@ def get_i18n_analyzer():
     except AttributeError:
         return getattr(analysis, 'standard')
 
+
 i18n_analyzer = LocalProxy(lambda: get_i18n_analyzer())
 
 
 @task(route='high.search')
-def reindex(obj):
-    model = obj.__class__
+def reindex(obj, id=None):
+    if id is not None:
+        model = db.resolve_model(obj)
+        obj = model.objects.get(pk=id)
+    else:  # TODO: Remove this branch in udata 2.0
+        warnings.warn(
+            'Document as task parameter is deprecated and will be removed in udata 2.0',
+            DeprecationWarning
+        )
+        model = obj.__class__
     adapter_class = adapter_catalog.get(model)
     timeout = current_app.config['ELASTICSEARCH_INDEX_TIMEOUT']
     if adapter_class.is_indexable(obj):
@@ -149,34 +160,42 @@ def reindex(obj):
 
 
 @task(route='high.search')
-def unindex(obj):
-    model = obj.__class__
+def unindex(obj, id=None):
+    if id is not None:
+        model = db.resolve_model(obj)
+    else:  # TODO: Remove this branch in udata 2.0
+        warnings.warn(
+            'Document as task parameter is deprecated and will be removed in udata 2.0',
+            DeprecationWarning
+        )
+        model = obj.__class__
+        id = obj.pk if isinstance(obj.pk, basestring) else str(obj.pk)
     adapter_class = adapter_catalog.get(model)
-    if adapter_class.exists(obj.id, using=es.client, index=es.index_name):
-        log.info('Unindexing %s (%s)', model.__name__, obj.id)
+    if adapter_class.exists(id, using=es.client, index=es.index_name):
+        log.info('Unindexing %s (%s)', model.__name__, id)
         try:
-            adapter = adapter_class.from_model(obj)
+            adapter = adapter_class(meta={'id': id})
             adapter.delete(
                 using=es.client,
                 index=es.index_name,
                 request_timeout=current_app.config['ELASTICSEARCH_INDEX_TIMEOUT'],
             )
         except Exception:
-            log.exception('Unable to unindex %s "%s"', model.__name__, str(obj.id))
+            log.exception('Unable to unindex %s "%s"', model.__name__, id)
     else:
-        log.info('Nothing to do for %s (%s)', model.__name__, obj.id)
+        log.info('Nothing to do for %s (%s)', model.__name__, id)
 
 
 def reindex_model_on_save(sender, document, **kwargs):
     '''(Re/Un)Index Mongo document on post_save'''
     if current_app.config.get('AUTO_INDEX'):
-        reindex.delay(document)
+        reindex.delay(*as_task_param(document))
 
 
 def unindex_model_on_delete(sender, document, **kwargs):
     '''Unindex Mongo document on post_delete'''
     if current_app.config.get('AUTO_INDEX'):
-        unindex.delay(document)
+        unindex.delay(*as_task_param(document))
 
 
 def register(adapter):