+++ udata/settings.py
@@ -431,7 +431,7 @@ class Testing(object):
     URLS_ALLOWED_TLDS = tld_set | set(['test'])
     URLS_ALLOW_PRIVATE = False
     # FakeSearch fields have to be declared here
-    SEARCH_FAKE_FIELDS = (
+    SEARCH_FAKESEARCHABLE_FIELDS = (
         'title^2',
         'description',
     )