+++ udata/tests/test_discussions.py
@@ -574,7 +574,7 @@ class DiscussionsMailsTest(FrontTestCase):
         )
 
         with capture_mails() as mails:
-            notify_new_discussion(discussion)
+            notify_new_discussion(discussion.id)
 
         # Should have sent one mail to the owner
         self.assertEqual(len(mails), 1)
@@ -594,7 +594,7 @@ class DiscussionsMailsTest(FrontTestCase):
         )
 
         with capture_mails() as mails:
-            notify_new_discussion_comment(discussion, message=new_message)
+            notify_new_discussion_comment(discussion.id, message=len(discussion.discussion) - 1)
 
         # Should have sent one mail to the owner and the other participants
         # and no mail to the commenter
@@ -619,7 +619,7 @@ class DiscussionsMailsTest(FrontTestCase):
         )
 
         with capture_mails() as mails:
-            notify_discussion_closed(discussion, message=closing_message)
+            notify_discussion_closed(discussion.id, message=len(discussion.discussion) - 1)
 
         # Should have sent one mail to each participant
         # and no mail to the closer